Clarify initial focus of project
Summary:
It's hard to write acceptance tests and unit tests when you don't have a clear picture
of what the program should be doing. I have a lot of ideas, but the program needs a
clear direction. This commit sets the stage to work initially on matching a list of
existing resources with a list of new resources for the purpose of batch replacing
the old with the new in a set of source files.

I still need to work up some kind of charter, so that the program's purpose isn't so
muddy. And I'm not sure I need all of these projects right now.